// src/Components/K/types/note.types.ts
export interface WorkspaceLink {
    workspaceId: number;
    workspaceName: string;
    workspaceItemId: number; // workspace_items.id
}